Add basic styling to the application UI
This also includes the concept of a base model type for 'standard pages'
which includes information used by common elements of the screen.

Controllers have been refactored and services created to populate this
information into the model base type, so that it is accessible anywhere.

Additionally a unit test for the login controller has been refactored to
make it a little less cluttered with things we don't care about.