WIP #AG57 - Add config transformations to Agiil.Web
This will create a Web.config file from a combination of the
Web.template.config file and one of a number of transformation
files.  This will ensure that the created file will always be
correct for the desired environment.

Note that the main Web.config file is now removed-from and
ignored in source control, because it is now generated on build.