Resolve #AG113 - Refactor database schema creation
This splits up the concept of creating a schema from the
concept of exporting a schema to DDL.  These two tasks are
now handled by two different interfaces.  Whilst in the
existing code they are fulfilled by the same class, this
doesn't need to be the case in all implementations.