Correct an anomaly in the sign-out logic
The sign-in and sign-out logic (usages of the IAuthenticationManager)
were not parallel.  One of them specifies the auth type and the other
does not list any auth types.  This corrects that mistake.

It was noticed whilst attempting to solve #AG66, although this work
has not helped in that ticket, it is still legitimate and should be
kept.