WIP #AG54 - Add a service which tests for the application dependencies
Right now this is just the database, but in future it could be "any
execution dependency" which should prevent the application from running
if it were missing.

Where the tests fail, the exact nature of the failure is recorded in the
application log file at either the WARN or FATAL levels (depending on
whether the problem can be fixed in maintenance mode or not).