WIP #89 - Add initial work on a QueryLanguage project
* README
* Some convenience scripts
*A copy of the ANTLR v4 binary

The reason I'm including the binary here is because it's not available on
any reliable package manager, such as NuGet (at least, not a recent
version).  The official download site could work but I don't want to be
at the mercy of them taking it down (say it becomes outdated, or they
change their URL scheme).