WIP #AG262 - Conversion to React complete
This completes the stand-still conversion of the view-ticket page
to React & Redux.  The page now renders in the same way as it did
before the JavaScript, but everything on the page is now handled
by React/Redux.

All that remains is to add the script to actually transition the state
of the current panels.