WIP #AG341 - Attempt to fix build failure with copy of assembly
The problem here is a conflict between two versions of CSF.Reflection,
v1.x and v3.x.  ZPT-Sharp needs 1.x and other usages need 3.x, and the two
are not compatible.

Because they are strong-named, I am able to copy BOTH versions into
the output directory (one with a different filename), and the assembly
bindings should sort out the ambiguity between them.

This isn't a great solution, and deserves a tech-debt ticket to undo, once
I'm able to only use a single version of the reflection library.