WIP #AG341 - Fix some NHibernate mappings, as demonstrated by tests
The persistence tests have shown that I have a problem with my one-to-many
mappings (as predicted).  The root cause was down to the fact that my
property mappings required the property to be writable, but where it comes
to collections, actually they don't need to be writable.

This was skipping those properties and meant that NHibernate was ignoring
them in the mapping.