Attempt to resolve #AG384 by altering OpenCover registration
This appears to be an issue whereby OpenCover (with the config I had been using in the
past) was broken by a change in the default AppVeyor image (it is now bundled
with .NET Framework 4.8).  That version of OpenCover I was using does not work with
the -register:user option under .NET Framework 4.8.  Apparently changing
to -register:Path64 fixes it.  More info available at:
* https://github.com/OpenCover/opencover/issues/911 (specific to AppVeyor)
* https://github.com/OpenCover/opencover/issues/914 (the root cause)