Fix crash error when login state is invalid
I think this was caused by having an existing browser cookie for the domain,
indicating that I am logged-in as a user which does not exist in the current
database.

This caused it to fail to retrieve a user from the DB and as a result crashed
because it had expected to find a user.